+++ apps/web-svelte/src/lib/components/VideoTranscriptPanel.svelte
@@ -451,11 +451,7 @@
     </div>
 
     {#if session.isAdmin}
-      <SubtitleSearchSurface
-        {videoId}
-        {onTracksChanged}
-        {onActiveTrackIdChange}
-      />
+      <SubtitleSearchSurface {videoId} />
     {/if}
   {/if}